feat(commands): §11e step 2 — registry + per-verb handler files (#264)
Replaces drain.ts's inline if/else dispatch with a registry lookup so
adding a new verb is one entry in registry.ts plus one file in
handlers/. Behavior-zero per kai+link locks (msg-1777932028252 /
msg-1777932030625):
  - same poll cadence
  - same ack contract
  - same unknown-skip semantics
  - no auth / intake churn
  - no new verbs (the two existing verbs become the first two registry
    entries — no scope creep)

New files:
  src/commands/types.ts                     — PendingCommand, CommandContext, CommandHandler
  src/commands/registry.ts                  — COMMAND_REGISTRY (context_sync, run_approve)
  src/commands/handlers/context-sync.ts     — relocated from drain.ts
  src/commands/handlers/run-approve.ts      — relocated from drain.ts

drain.ts is now ~80 LOC (down from ~200): poll + error tracking + dispatch
via registry with the existing unknown-skip ack fallback. Net diff +206/-136.

Closes §11e step 2 of reflectt-cloud/docs/COMMAND_PLUGIN_SKILL_ARCHITECTURE.md.
Step 3 (routes/canvas extraction) and step 4 (routes/chat extraction) are
explicitly out of scope for this PR.

Proof:
  - typecheck green
  - 244/244 test files pass (2647 tests, 1 skipped)
  - canonical staging command-queue smoke planned post-merge (poll, ack-complete, ack-fail)
